+++ tilings/algorithms/__init__.py
@@ -1,2 +1,4 @@
+from .factor import (Factor, FactorWithInterleaving,
+                     FactorWithMonotoneInterleaving)
 from .fusion import Fusion
 from .row_col_separation import RowColSeparation